requirements.txt
langchain
langchain-openai
langgraph
pydantic
python-dotenv
matplotlib
